requirements.txt
twisted
pyyaml
sphinx
argparse
pbr
setuptools
